potential fix for get_message lockup
If a command is sent in a dead chat (just the bot and the client) and the listener doesn't capture the response to the command the listener thread will lockup when calling get_messages on the "resp = sock.recv(2048).decode('utf-8')" this will cause the main thread to lockup waiting for the listener thread even if "kill_switch" has been set with "kill_switch.set()". By passing "kill_switch" to "get_messages" and using it as a condition for the while loop with an outer while 1 loop this issue may be able to be fixed.